fix a bunch of small vector+matrix things
svn path=/trunk/yarp2/; revision=4932

+++ src/libYARP_dev/src/RemoteControlBoard.cpp
@@ -1766,7 +1766,7 @@ public:
     virtual bool getEncoders(double *encs) {
         Vector *v = state_buffer.read(true);
         if (v != NULL) {
-            ACE_ASSERT (v->size() == ((unsigned int)nj));
+            ACE_ASSERT (v->size() == nj);
             ACE_OS::memcpy (encs, &(v->operator [](0)), sizeof(double)*nj);
             return true;
         }
